HashFilter: A simple append-only hash-based value recognisor
After much, *much* deliberation about how to store incompressible paths,
this seems to be what I'm settling on, for now.  A simple list of u128
hashes, appended to a file and stored in a HashSet<u128>.

A scalable bloom filter was considered, but this meets my desire for
being compact reasonably well while also being simple, fast,
IO-efficient, and relatively robust.