Update deps
Switch to wfd for folder dialogs, and - for the time being -
tinyfiledialogs for error modals.  Its use of String for file paths
makes it unsuitable for folder selection.

Remove my DPI-awareness hack now web-view handles that.  This seems to
have adjusted some base sizes, so also tweak style and window size to
compensate.